Retrieve the categories of a country-specific variable in LIS/LWS for a given country.

Retrieve the categories of a country-specific variable in LIS/LWS for a given country.

Usage

get_country_specifc_categories(
  variable,
  iso2,
  from = NULL,
  to = NULL,
  lws = FALSE,
  n_categories = FALSE
)

Arguments

variable

A unit-length character vector containing a LIS/LWS country-specific (with the "_c" suffix) variable.

iso2

A character vector with a valid iso2 code for countries present in LIS/LWS.

from

A numeric value representing the year (inclusive) after which the LIS/LWS datasets should be considered.

to

A numeric value representing the year (inclusive) up to which the LIS/LWS datasets should be considered.

lws

A logical value, that guides the tool to search in the LIS or LWS database. The argument is FALSE by default, taking LIS as the databse to be investigated if nothing is specified.

n_categories

A logical value indicating whether to output the number of categories of a single country-specific variable, across the entire time series for a given country.

Value

A list made of character vectors. If n_categories = TRUE, then a list with a numeric vector.

Examples

# In years where no data is recorded for a given variable, it is automatically hidden from the output
variable_exists(variable = "health_c", iso2 = "it")
#> $it
#>  1977  1978  1979  1980  1981  1982  1983  1984  1986  1987  1989  1991  1993 
#>  "No"  "No"  "No"  "No"  "No"  "No"  "No"  "No"  "No"  "No"  "No"  "No"  "No" 
#>  1995  1998  2000  2002  2004  2006  2008  2010  2012  2014  2016  2020 
#> "Yes"  "No"  "No"  "No"  "No" "Yes" "Yes" "Yes"  "No"  "No"  "No"  "No" 
#> 
get_country_specifc_categories(variable = "health_c", iso2 = "it", from = 1995, to = 2020) 
#> Error in get_country_specifc_categories(variable = "health_c", iso2 = "it",     from = 1995, to = 2020): object 'final' not found

# To retrieve information on LWS datasets
get_country_specifc_categories(variable = "bus1_c", iso2 = "fi", lws = TRUE)
#> Error in get_country_specifc_categories(variable = "bus1_c", iso2 = "fi",     lws = TRUE): object 'final' not found

# Using the `n_categories` argument
get_country_specifc_categories(variable = "region_c", iso2 = "es", n_categories = TRUE)
#> Error in get_country_specifc_categories(variable = "region_c", iso2 = "es",     n_categories = TRUE): object 'final' not found

# To use this function acroos multiples countries one could make use of the `purrr::map()` function 
purrr::map(lissyrtools::show_countries_lws(), ~get_country_specifc_categories(variable = "bus1_c", iso2 = .x, lws = TRUE , n_categories = TRUE))
#> Warning: The selected variable: bus1_c, does not have values other than zeros or missings for the selectd years.
#> Error in purrr::map(lissyrtools::show_countries_lws(), ~get_country_specifc_categories(variable = "bus1_c",     iso2 = .x, lws = TRUE, n_categories = TRUE)): ℹ In index: 1.
#> ℹ With name: Australia.
#> Caused by error in `get_country_specifc_categories()`:
#> ! object 'final' not found