Inspect the Labels of LIS and LWS variables

Inspect the Labels of LIS and LWS variables

Usage

get_vars_label(vars = NULL)

Arguments

vars

A character vector containing LIS/LWS variables or the output list from lissyuse.

Value

A character vector with the corresponding labels for the selected variables.

Examples

# 1) Without any argument:
get_vars_label()
#> Error in deframe(data_vars_labels): could not find function "deframe"

# 2) Using with the outputed list from lissyuse:
lis_datasets <- lissyuse(data = c("uk"), vars = c("hpub_i","hpub_u", "hi42", "hi421", "hi422", "hi43"), from = 2016)
#> Error in import_sample_datasets_to_lissyuse(data, lws): Invalid value for 'data'. When lws = FALSE, only the sample datasets stored in this package can be loaded in the following format: 'us', 'it', 'mx', 'us16', 'it14', 'mx18'.
get_vars_label(vars = lis_datasets) # ???? quotes, no quotes ? 
#> Error: object 'lis_datasets' not found

# 3) Using a character vector with LIS/LWS variables:
get_vars_label(vars = c("fyft", "basb", "hxremit", "bafi1_c", "pasodc"))
#> Error in deframe(data_vars_labels): could not find function "deframe"