Plot an Indicator.

Computes an indicator and plots it for one or multiple LIS or LWS datasets.

plot_indicator(
  list_datasets,
  variable,
  indicator,
  plot_theme = NULL,
  ratio = NULL,
  na.rm = FALSE,
  ...
)

Arguments

list_datasets

A list of LIS or LWS datasets.

variable

A character string indicating the aggregate for which the indicator needs to be plotted.

indicator

A character string indicating the type of indicator to be computed. Currently the function supports only 'mean', 'median', 'ratio' and 'gini'.

ratio

Only used in the computation of 'ratio' indicator. A vector of two numeric values between 0 and 1. Defines the percentiles in the numerator and denominator respectively. E.g. (0.9, 0.1) computes the 90/10 ratio.

na.rm

A boolean. Indicates if NAs should be ignored. Defaults to FALSE.

theme

A ggplot2 theme for the graph.