Read the format of files argument

Use 'read_file_name_format' instead.

Detects the format of the 'file' argument passed to 'read_lissy_files_locally()' and returns it in a 'two digit country, two digit year, one digit level (optional)' format. The names of the files this function can read need to indicate the country ('c'), year ('y') and if it's a household or person level dataset ('l'). In addition, it can optionally indicate if it's a LIS, LWS or ERFLIS dataset ('d'). This information can be in the following formats:

'ccyyl' Two digit country, two digit year, one digit level. E.g. "fr84h"

'ccyyyyl' Two digit country, four digit year, one digit household/person ('h' or 'p'). E.g. "fr1984h"

'ccyydl' Two digit country, two digit year, one digit indicating LIS, LWS or ERFLIS (i/w/e/r), one digit household/person. E.g. "fr84ih"

'ccyyyydl' Two digit country, two digit year,one digit indicating LIS, LWS or ERFLIS (i/w/e/r), one digit household/person. E.g. "fr1984ih"

It can contain any printable characters after the pattern (e.g. 'fr84ih_new').

@return A file name with the format 'ccyydl' or 'ccyyl' depending if information on the database is provided or not.

Usage

read_format_names(file_name)