Verify if a given variable has a note, for a selected group of countries, in LIS or LWS databases.

Verify if a given variable has a note, for a selected group of countries, in LIS or LWS databases.

Usage

variable_has_note(variable, iso2, lws = FALSE)

Arguments

variable

A unit-length character vector containing specified LIS/LWS variables.

iso2

A character vector with valid iso2 codes of countries present in LIS/LWS

lws

A logical value, that guides the tool to search in the LIS or LWS database. The argument is FALSE by defualt, taking LIS as the databse to invetigated if nothing is specified.

Value

A list, made of character vectors. Each elements corresponds to a country in LIS or LWS databases.

Examples

variable_has_note(variable = "area_c", iso2 = "br")
#> Error in map(iso2, process_country): could not find function "map"
variable_has_note(variable = "basb", iso2 = c("fr", "de", "us", "uk"), lws = TRUE)
#> Error in map(iso2, process_country): could not find function "map"